requirements.txt
alembic==1.18.1
anitya_schema==2.3.0
annotated-doc==0.0.4
annotated-types==0.7.0
antiorm==1.2.1
anyio==4.12.1
app==0.0.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==25.4.0
Automat==25.4.16
automium==0.2.6
automium_web==0.1.1
backend==0.2.4.1
base==0.0.0
bcrypt==4.0.1
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorclass==2.2.2
constantly==23.10.4
crochet==2.1.1
cryptography==46.0.3
db==0.1.1
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
ecdsa==0.19.1
fastapi==0.128.0
fedora_messaging==3.9.0
fedora_messaging_the_new_hotness_schema==1.3.0
Flask==3.1.2
gssapi==1.10.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
koji==1.35.3
legacy==0.1.7
Mako==1.3.10
MarkupSafe==3.0.3
MyApplication==0.1.0
packaging==25.0
passlib==1.7.4
pika==1.3.2
pip-upgrader==1.4.15
pluggy==1.6.0
psycopg2-binary==2.9.11
pwdlib==0.3.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pytest==9.0.2
python-bugzilla==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
PyYAML==6.0.3
redis==5.3.1
referencing==0.37.0
requests==2.32.5
requests-gssapi==1.4.0
rpds-py==0.30.0
rsa==4.9.1
service-identity==24.2.0
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
terminaltables==3.1.10
the-new-hotness==1.3.0
tomli==2.4.0
Twisted==25.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.5
wrapt==2.0.1
zope.interface==8.2
